Merge pull request #91 from kitsunet/datetimeinterface
[TASK] Use DateTimeInterface in Flow

As Flow requires PHP5.5 we should annotate and use the
``\DateTimeInterface`` to allow usage of the ``\DateTimeImmutable``
in all places appropriate.

Since PHP 5.5 there is a DateTimeImmutable that will always
return clones of itself on any mutating method call. Using that
for ``Now`` makes sense. Additionally code is adapted to accept
instances of \DateTimeInterface which is needed for Now.

Resolves: FLOW-140