Merge pull request #170 from kdambekalns/keep-response-status
[BUGFIX] Keep status code from module response on redirect

The response of a module loses the status code if it contains a Location
header. This is caused by the ModuleController redirecting to the found
location but ignoring any status code that may have been set. This means
that a module will only create 303 redirects.

This change takes the status code of a module response and applies it to
the generated redirect, fixing this behavior.