Merge pull request #178 from robertlemke/flow-improvemissinghomepagelogging
[TASK] Improve logging for "no homepage found" situations

When the Node Route Part Handler wasn't able to find a site node,
the exception now contains information about if at least a domain
could be detected and if so, which one matched.

Additionally, matchValue() will now also log an exception if the
request path was empty (that is, the user requested the homepage) and
not only for sub pages.