Merge pull request #211 from robertlemke/workspace-management-fixes
[BUGFIX] Broken workspace management backend module

This fixes a couple of issues with the new workspace management features
and improves the user interface in the workspace review screen.

Editing the owner of a workspace resulted in a fatal error due to
a wrong method signature. Ownership can now be transferred as intended.

The rebase and redirect action (which can conveniently be triggered
through an edit button in the review screen) complained about a wrong
"$this" in a fatal error. Removed $this and now that works.

At the bottom of the review workspace changes screen a small help
message explained what would happen when a user presses one of the
buttons. Removed that fine print and now let the buttons speak for
themselves.