Merge pull request #216 from robertlemke/neos-1683-fix-workspace-name-suffix
[BUGFIX] Workspace identifier suffixes are not random (enough)

This solves an issue with the automatically generated suffix for
names of workpsaces created through the Workspace Management module.

Previously, workspace suffixes (like "-7603") were identical or very
close to previously generated names, because they were derived from
the first four characters of the output of uniqid(), which is based
on microtime.

The new solution converts the current microtime into an alpha-numeric
string and takes the last 5 characters, which change very often.

Resolves: NEOS-1683