Merge pull request #107 from albe/valueobject_injection
[BUGFIX] Ignore injected properties for value hash generation

When injecting properties into a value object, an exception was thrown that
a closure is tried to be serialized. This was due to the hash generation
in PersistenceMagicAspect only skipping properties that are annotated as
transient.

This change makes the value hash generation resort to the class schema instead
of directly iterating all properties, since the class schema is already focused
on persistence relevant properties and hence contains no injected properties.