Merge pull request #138 from kitsunet/flock-lock-refactor
[BUGFIX] FlockLockStrategy should clean up lock files

The FlockLockStrategy creates files to apply the lock on.
These files reside in the temporary folder but are never cleaned
on releasing the Lock that means the amount of files in this folder
will increase over time unless the folder is cleared manually.

Additionally cleans the code a bit and reduces chance of race
conditions while creating the lock.