Merge pull request #153 from kdambekalns/php-53-compat
[BUGFIX] Fix PHP 5.3 compatibility issues

The 2.3 branch should be compatible to PHP 5.3, but since we did not have automated
testing in place on that version of PHP, some incompatibilties have crept in.

This fixes those issues and enables the unit and functional tests to be run on PHP 5.3
on Travis CI.

Also the Behat tests for Flow are run with this, since there is no reason they shouldn't.