Merge pull request #269 from robertlemke/workspace-side-by-side
FEATURE: Side-by-side comparison of changes in Workspace Management

This change adds a side-by-side view of changes of a given workspace
to the "Review" view of the Workspace Management module. It highlights
inline text changes and displays thumbnails of changed assets.

This new feature relies on a third party diff library which has been
re-packaged, moved to a proper namespace, made Composer auto loading
compatible and been made compatible with PHP 7. It is from now on
part of the Neos distribution.

NEOS-956 #close