Merge pull request #179 from neos/do-not-unlink-lockfiles
BUGFIX: Don't unlink lock file

Unlinking the lock file again causes race conditions which the
locking should prevent in the first place. Until we found out a
solid way to do this we need to keep those files even though
this is not optimal due to the high amount of inodes those lock
files use up.