Merge pull request #283 from aertmann/site-import-broken
BUGFIX: Do not automatically persist thumbnails for new asset objects

Prevents issues with database constraints caused by new assets being added to persistence, since
the asset would be added through the thumbnail as a related object as well.

Due to the automatic creation of thumbnails on asset creation, this caused problems for the site import.

NEOS-1671