Merge pull request #188 from dfeyer/task-ressourcemanager
TASK: Add method to get public URI by path

This change introduces two new methods in the ResourceManager.

The method ``getPackageAndPathByPublicPath`` returns an array with package and
path. This method is useful to avoid duplication of this logic. The second method ``getPublicPackageResourceUriByPath`` is syntactic sugar.

This change also adapts the ``ResourceViewHelper`` to use the new methods.