Merge pull request #109 from albe/cyclic_validator_chain
[BUGFIX] Validator chain is built correctly for cyclic relations

For cases where an entity A references entity B and B in turn
references A (as first property), the validator chain would not
validate the reference to A inside B, even though those might be
different instances.
This is due to the newly created ObjectValidator for B not being added
to the CollectionValidator inside A - so when B again visits A, the
returned validator collection is empty and hence not added to the
validator chain.

This change fixes that, by adding the new ObjectValidator to the chain
early, and only removing it later on, if it is still empty after all
properties being checked. This could potentially lead to a deeper
validator chain than would be necessary (cycle of empty ObjectValidators),
but should not break anything or degrade performance.

Fixes: FLOW-352