Merge pull request #288 from kitsunet/node-repair-default-values
TASK: node:repair doesn't mix internal node properties

The node:repair command would currently apply any configured default
value as node property value if the property did not exist. This is
wrong for properties prefixed with an underscore as we use this
prefix for internal properties.
This is now changed so that any property starting with underscore
is ignored.

As this prefix is currently implicit knowledge a small note was added
to the documentation, stating that property names should not start
with underscores as we use them internally.

NEOS-1721 #close