Merge pull request #191 from kszyma/patch-1
BUGFIX: Unlocking site breaks on Windows

In the method "unlockSite" the order for unlocking a file is wrong. Under Windows you first have to flock/fclose and then unlink (otherwise the lockfile is locked by OS and can not be removed). Then you get in Context Production the Errormessage "Site is currently locked, exiting.". The above reordering will fix this.