Merge pull request #311 from kitsunet/workspace-name-inconsistency
BUGFIX: User workspace names should be deterministic

This fixes some inconsistencies with the naming of user workspaces,
which resulted in duplicate workspaces being created for users with
special characters in their username.

Additionally a migration that can correctly set the owner for users
with usernames containing special characters was added.

NEOS-1740 #close Fixes the issue with a new migration