Merge pull request #202 from kitsunet/fix-flow-431-12
BUGFIX: Default constructor arguments only autowired if needed

So far all constructor arguments were treated as autowired unless
configured otherwise. This can result in problems and one example
is seen with the constructor of ArrayObject starting from PHP 7.0.2.

Unfortunately PHP does not allow Reflection of built-in functions
arguments. This includes methods of built-in classes like ArrayObject.
This results in our ReflectionService to report null as default for
those arguments which will then be autowired. Since PHP 7.0.2 all three
arguments of the ArrayObject constructor are reported correctly by
Reflection but due to our autowiring the third argument will be given
as null in the proxy class which results in a fatal error.

The solution is to mark configured arguments that use the default or null
to be not autowired and do not give them to the original constructor if
they are not followed by an autowired argument.

FLOW-431 #close Fixes the issue by changing the way we apply default values