Merge pull request #200 from albe/proxy_fix
BUGFIX: AOP Proxies properly check method existence before upward delegation

The usage of method_exists() checks inside the proxy class code was wrong by checking
the existence on the parent class of $this, which might in some circumstances not be
the current proxy class (ie. when checking method was originally called from a doctrine proxy).

This change fixes the method_exists calls by using the get_parent_class() without
parameter. Also, the method_exists check is added to the remaining method checks, that
currently only rely on is_callable, which in turn is always true on hierarchies that implement
__call magic method.