Merge pull request #131 from kitsunet/pm-refactor-nonbreaking
!!!TASK: Refactor package management

This is a major refactoring of the package management.
With this change the structure of the PackageStates file
is overhauled now using the composer key for consistency to
avoid problems if namespace or autoload information changes
as that would change the (Flow) package key.

Additionally makes the resolution mechanism for Flow package
keys more robust and stores additional metadata in the
PackageStates.php to avoid loading all composer manifests
at runtime.
Furthermore the package loading order generation was improved by
using a more sophisticated algorithm to get the order right.

All the changes together lead to a noticeable speed improvement
in runtime situations.

Deactivating packages now results in the package being moved out
of regular package paths into an Inactive directory.

This change is marked breaking due to the following:

New packages are only automatically picked up if they were
installed via composer or created through the Flow package
management. In all other cases you need to run the:
package:rescan command to pick new package up.

Some @api classes and methods were deprecated and will be removed
or changed in the next major Flow version.

Some newly added methods of PackageManager and Package are used
in the Flow core now, so if someone would reimplement both
according to interface it would not work with Flow, but that
is already the case without this change, so it shouldn't be an issue.

The whole MetaData part of the package management is no longer
used and will be removed in the next major version.