Merge pull request #215 from kitsunet/deprecate-early-logger
TASK: Deprecate EarlyLogger

The ``EarlyLogger`` was explicitly meant for the ``PackageManager``
which is not using it anymore. Therefore the EarlyLogger is deprecated
and bound for removal in the next major Flow version.

To underline the fact that it was never meant for user land code the
``@ap``i annotations were removed.