Merge pull request #239 from mkeitsch/fix-flow-305
BUGFIX: 'Content-Type' (automatic) header fails in virtual browser

With adding the `Content-Type` header to the automatic headers of a virtual browser, the request fails every time.

The given value is cast to an array by the `Http\Headers::set()` method. When setting this header in a `Header` (!) instance (`Http\AbstractMessage::setHeader()`) of the request, a string is expected especially for `Content-Type`, but an array is given.

`BrowserTest` extended especially for this header field.

FLOW-305 #close