Merge pull request #217 from kdambekalns/gerrit-41942
BUGFIX: Evaluate property conditions without security checks

The PropertyConditionGenerator for entity privileges allows the use of
global objects from the configured global context. If those in turn
may be secured, the system runs into an endless loop.

To avoid this, the fetching of the value for the operand is done without
security checks after this change.