Merge pull request #392 from robertlemke/missing-site-node-in-nodes-service
BUGFIX: Translate process is broken for dimensions without fallbacks

This fixes an issue for sites with single or multiple dimensions which
have no fallback defined. When trying to translate content from an
existing dimension to a variant which does not exist yet, the
``NodesController`` will fail with a fatal error due to the missing
Site Node in the respective dimension.

The solution in this change works for the given situation. A more
generic approach may be implemented as part of
pull request 156.

NEOS-1786 #close