Merge pull request #300 from kitsunet/fix-configuration-caching
TASK: Fix Configuration cache usage

The Objects configuration was never checked against the loaded
configuration cache but still saved sometimes. If a specific configuration type
was saved depended on the fact that it was loaded at the time of writing the
configuration cache. We now make sure that Objects configurations are also used
if they are cached and that we load all configuration types before writing the
cache.