Merge pull request #304 from dfeyer/bugfix-request-create
BUGFIX: If query is null the condition break

This change casts the query (string) to string before checking if we need to append it.
Without this change the condition to build the request URI is wrong when the query is null
and the URI contains a question mark without request arguments.