Merge pull request #416 from aertmann/inspector-breadcrumb
TASK: Use node label in inspector breadcrumb

Instead of using the node type name the label of the node is used instead. This helps avoid confusing the breadcrumb in the inspector from a node type selector.

Additionally the label is now updated using the server side rendered label in the inspector breadcrumb, context structure tree and node tree.

NEOS-1141 #comment replaced the node type name with the node label, which should help a little