Merge pull request #317 from kitsunet/deprecate-property-mapping-config
TASK: Get rid of PropertyMappingConfigurationBuilder

The ``PropertyMappingConfigurationBuilder`` its superfluous and contains
only one method that can easily be integrated into the ``PropertyMapper`` as
you will need that anyway to do something with the configuration.

Therefore ``PropertyMappingConfigurationBuilder::build()`` is deprecated
and fully replaced by ``PropertyMapper::buildPropertyMappingConfiguration()``.
The ``PropertyMappingConfigurationBuilder`` will be removed for the next
major Flow version.