Merge pull request #454 from kdambekalns/compare-workspaces-by-name
TASK: Find workspacePosition by comparing names

Two places use workspaces in array_search to find find the positions of a
workspace. This fails if two instances of the same workspace are used,
and even though this should not happen, the error is not "needed". When
using the name to compare the workspaces, everything works fine (as is to
be expected, since the name is the identifier).