Merge pull request #462 from mficzel/bugfix/freeIndexSpaceInsteadOfRenumbering
BUGFIX: Free index space at target position if no free index space is available

If a node is inserted at a given position between nodes and no free sorting index is available, the sortindices on
that level are renumbered. The previous code for that could lead to unexpected node reordering and sortingindex
value escalation if workspaces or dimensions were used.

The following steps reproduce the error:

    In dimension A create nodes between other nodes until there are no free sortindices available
    Create a variant of those nodes in dimension B
    In dimension B add a new node in a place where no free sort index is available

Since the previous code is only repositioning one item of a given index and does not take workspaces and dimensions
into account this results in the following unwanted effects:

    Unwanted reordering of the nodes in dimension A
    In dimension B two nodes with identical sortingindex occur which makes the order of the nodes random
    If this is repeated multiple times the sorting indices in dimension A escalate quickly to very high values

This patch resolves this behavior by freeing index space at the target position instead of renumbering the
whole level by modifying all nodes on the given path and incrementing all sort indices above the reference position
a consistent behavior across workspaces and dimensions is ensured.