Merge pull request #332 from kitsunet/package-rescan-fix
BUGFIX: package:rescan must happen before booting

If the user issued a ``typo3.flow:package:rescan`` command this should
happen very early in the compiletime boot sequence as otherwise the
ObjectManager will try to load a non existing package if the rescan was
issued after a package was removed from the installation.

With this change package rescans are possible in all package state
situations in case you manipulate the packages without using composer
or our ``PackageManager``.

Additionally adds a bunch of use statements to the ``Bootstrap``.