Merge pull request #479 from radmiraal/bugfix/getidentifier-on-nonobject
BUGFIX: Call to getIdentifier on non object in usage references

In the usages overview for assets in the media module a list is rendered
with a path to the node where an asset is still used. This leads to a
call to getIdentifier() on a non object if the parent document node
is removed in the live / current users workspace.

As we can not display the title of the document in the other users workspace
this change leaves out the breadcrumb to the node in the overview.