Merge pull request #498 from kdambekalns/neos-1633
BUGFIX: Support empty dimension values for default fallbacks

Implement support for empty dimension values (for any number of dimensions) of a node variant.
This allows to have nodes like ``/sites`` that will always be found, independent from the dimensions
of the context.

Another case is the transition to new dimensions that will be much easier because existing nodes
are already visible.

NEOS-1633 #close