Merge pull request #350 from kdambekalns/bugfix_ishas_access
BUGFIX: Fix ObjectAccessorNode::getPropertyPath for is/has access in PHP7

This fixes the try/catch for undefined Fluid variables which start with "is" or "has".
Currently, exceptions here are not being caught because an `\Error` is being thrown.
Since non-callable method call should not be attempted in the first place, this change
adds a is_callable check. Therefore a try/catch block is no longer necessary, as any exception thrown inside the callable method should bubble up.

Specific example: a fluid section/partial with an optional argument called "isSomething"

- if rendered while defining isSomething in the arguments, there's no problem
- if rendered without isSomething in the arguments (implicit falsy), we then get
  `Call to undefined method TYPO3\Fluid\Core\ViewHelper\TemplateVariableContainer::isSomething()`

Related to: #108, #343, #348, whythecode/flow-development-collection#1