Merge pull request #347 from DavidSporer/documentation-many-resources
TASK: Document handling a large number of Resources

If a project contains lots of resources, resource:publish might fail if the maximum number of folders is being reached when publishing resources because no more folders can be created.
This part describes how to use the subdivideHashPathSegment target option in order to circumvent this error.