Merge pull request #362 from mficzel/bugfix/avoidTypeCastingErrosIfComplexValueValidationFails
BUGFIX: Avoid type-casting errors if the validation of complex types fails

If an error is created in some cases the value was added to the result-message without checking for the type beforehand, that could lead to php errors if the value could not be casted to a string.

This change centralizes the rendering of the value in a 'renderValue' - method.