Merge pull request #627 from robertlemke/feature-noderepair-removeinvaliddimensionnodes
FEATURE: node:repair: remove nodes with invalid dimension values

This change contains an improvement for node:repair which runs an
additional check for nodes which have dimension values which are not
allowed according to the current dimension configuration.

These nodes can be either removed or the user can decide to migrate them
by other means (for example a node migration).