Merge pull request #130 from bwaidelich/requestPattern-options
TASK: New configuration syntax for request patterns & firewall filters

The configuration format for authentication request patterns is
not very flexible and inconsistent with other settings.

With this change the syntax is changed from:
```yaml
  'SomeAuthenticationProvider':
    # ...
    requestPatterns:
      'controllerObjectName': 'Some\Package\AdministrationArea\.*'
```

To a more generic syntax:

```yaml
  'SomeAuthenticationProvider':
    # ...
    requestPatterns:
      'Some.Package:AdministrationArea':
        pattern: 'ControllerObjectName'
        patternOptions:
          'controllerObjectNamePattern': 'Some\Package\AdministrationArea\.*'
```

The new syntax seems more cumbersome but it allows for advanced request
pattern options. Besides the new level allows to attach *multiple*
patterns of one type to a provider.
This also allows to *unset* previously configured patterns:

```yaml
  'SomeAuthenticationProvider':
    requestPatterns:
      'Some.Package:AdministrationArea': ~
```

The configuration syntax for request patterns for the firewall has
been adjusted accordingly::

```yaml
    TYPO3:
      Flow:
        security:
          firewall:
            filters:
              'Some.Package:BlockedUris':
                pattern: 'Uri'
                patternOptions:
                  'uriPattern': '\/some\/uri\/.*'
                interceptor: AccessDeny
```

This is not a breaking change as the old syntax is still supported.

But as the new syntax is recommended, this comes with a corresponding
core migration that you can execute as follows:

```
  /flow core:migrate --version 20151113161300 --verbose
```

FLOW-412 #close