Merge pull request #409 from albe/trusted_proxies_component
FEATURE: Add "Trusted Proxies" as a HTTP component

This change introduces a HTTP component for so called "Trusted Proxies" that is registered
at the start of the HTTP chain.

It is configurated via the setting `TYPO3.Flow.http.trustedProxies.proxies`, which is a list of
IP address ranges that are trusted to provide header overrides for some import request meta
informations like the host, port, protocol and client IP.

    proxies:
      - '216.246.40.0/24'
      - '216.246.100.0/24'

Also the header names that are trusted for those informations can be set via the configuration
option `TYPO3.Flow.http.trustedProxies.headers` and default to this:

    headers:
      clientIp: 'Client-Ip, X-Forwarded-For, X-Forwarded, X-Cluster-Client-Ip, Forwarded-For, Forwarded'
      host: 'X-Forwarded-Host'
      port: 'X-Forwarded-Port'
      proto: 'X-Forwarded-Proto'

The `clientIp` setting is chosen for backwards compatibility of the `getClientIpAddress` method.
This method will now resolve to the first IP address that is not a trusted proxy inside the first of the configured headers that exists in order from left to right.
This setting should be changed to only a single trusted header instead.

By default all IP addresses are trusted to stay mostly backwards compatible, but a more
safe setting would be to trust no proxies by default:

    proxies: []

FLOW-414 #close