Merge pull request #442 from kitsunet/fix-doctrine-controller
BUGFIX: Doctrine controller should allow empty host for database

The doctrine controller checked for correct configuration of a database
before calling the doctrine methods. As the drivers all have different
options the ``host`` is a not a conclusive check. The only thing that
can be checked is if a ``driver`` was set.