Merge pull request #706 from robertlemke/bugfix-noderepaircausesinconsistentidentifiers
BUGFIX: Workspace publishing fails after node:repair

This change addresses an issue which results in a fatal error caused
by foreign key constraints when a user tries to publish her changes
to another workspace. The root cause is that the task in node:repair
which fixes "unstable" node identifiers of auto-created child nodes
only changes identifiers of nodes in a specific workspace (by default
the "live" workspace) and by that irreversibly disconnects corresponding
nodes in other workspaces.

The fix consists of two parts: First, the "create missing child nodes"
task of node:repair is modified so that node identifiers are always
changed accross all workspaces. Second, there is a new node:repair
task which detects and fixes inconsistencies caused by this bug.

If you experience the symptom described earlier, simply run a
node:repair to fix the inconsistencies.

The steps to reproduce the bug based on the Neos Demo site (as of
August 1st, 2016) are:

1. Log in to the backend and modify the title of the page "Forms". Don't
   publish the change yet.
2. From the command line call node:repair and see that child node
   identifiers are changed.
3. Try to publish the changes to the live workspace.

The UI will report an error which is caused by a failed SQL update.
With this patch applied, the changes should be able to publish without
any errors.