Merge pull request #853 from kitsunet/bugfix/peg-parses-no-proxy
BUGFIX: Peg generated parsers are excluded from proxy building

The generated parsers ``AbstractParser``, ``EelParser`` and ``FizzleParser``
are excluded from proxy building as the include statement and the general
structure of these make them unsuitable for proxy-ing. This was only a bug
before if an AOP advice targeted the mentioned classes, now this cannot
happen anymore.

Fixes: #796