Merge pull request #854 from kitsunet/bugfix/proxy-code-newline
BUGFIX: Proxy Compiler adds newline after end of original code

As any original class might end on a comment line without newline we need to
add a newline as part of the proxy building to avoid putting the proxy
namespace declaration within a comment. Additionally adds a marker comment
declaring the beginning of the proxy code.

Fixes: #779