Merge pull request #1410 from mficzel/task/makeFusionParserAspectable
BUGFIX: Make fusion-parser aspectable

The fusion parser cloned itself previously to handle includes but to keep the defined namespaces.
That made the use of aspects impossible since they were only applied on the initially created instance of the parser.

This change fixes this by using a new instance of the parser for each include and transfers the namespaces explicitly to the new parser.