Merge pull request #886 from jhinzmann/task-increase-max-length-of-pdo-cache-entries
TASK: Increase max-length for cache entries in PdoBackend

When not using SQLite for the `PdoBackend` and manually creating
the cache tables by using the `DDL.sql`, the content column of the
cache table with type `TEXT` could easily lead to cropped
caching entries because of the length restriction.
Using `MEDIUMTEXT` practically prevents this issue.

See also: neos/flow-development-collection#884 and
neos/flow-development-collection#885