Merge pull request #1486 from daniellienert/bugfix/only-handle-assetusages-in-site-nodes
BUGFIX: NodePropertiesStrategy in Neos should only handle site nodes

Currently the AssetUsageInNodePropertiesStrategy handles
all asset references in all nodes in the node data table.
The handling than is specific for nodes under the `/site` root node
and fails when a node references an asset that is located under another root.

This fixes the issue by only selecting site nodes.

**What I did**
Filter the nodes by site root.

**How to verify it**
Add an asset reference to a node that is not under sites.
The edit view of an asset fails.