Merge pull request #1478 from mficzel/task/addIndexForParentpathField
BUGFIX: Add mysql-index for the field `parentpath` in the `nodeData`-table

The find operation is currently very expensive because of an unoptimized LIKE statement.
This index will to optimize the performance for find operations especially in larger cr-instances

The size of the index is a compromise between space and an assumption of the normal length of the nodePath.