Merge pull request #1560 from mficzel/feature/evaluateAtIfInFusionFalysOrTruthy
FEATURE: Evaluate `@if` in fusion as falsy or truthy values

The behavior of `@if` is altered to make the distinction between falsy or truthy values
and no longer check for an exactly `false` value in the condition-expression. For the
distinction the php rules for casting to boolean are applied.

Examples for falsy-values that are now detected in `@if`:
- empty array
- number zero
- null
- empty string