Merge pull request #1601 from neos/bugfix-1580
BUGFIX: Reset broken properties to array

If the content of the `properties` property cannot be decoded from JSON
correctly, it will be `null`. This leads to errors when any operation is done
that expects it to always be an array.

This change adds a PostLoad Doctrine lifecycle method to reset `properties`
to an empty array if it is `null` after reconstitution.

Fixes issue #1580