Merge pull request #1618 from robertlemke/bugfix-1617
BUGFIX: Avoid orphaned content nodes when calling publishNode()

This changes an issue with using the `PublishingService::publishNode()`
which can result in an inconsistent structure in a user's workspace.

This change also changes the behavior of `PublishingService::discardNode()`
which now will also discard content of a given document node to protect
consistency.

Document nodes and their content must be published or discarded together
in order to protect against inconsistencies when the document node is
moved or removed in one of the base workspaces.

Fixes #1617