Merge pull request #1004 from beheist/feature-add-package-type-to-kickstarter
FEATURE: Allow setting the package type when kickstarting a package

You can now do `./flow kickstart:package Foo.Bar --packageType neos-plugin`, like in the PackageCommandController.

A question regarding the docs: Is the "Command Reference" part of the documentation autogenerated? If so, it seems to be outdated, because there are still some mentions of typo3 in there.