Merge pull request #1495 from mficzel/fusionDsls
FEATURE: Add extension point for domain specific languages to fusion

DSLs are implemented for fusion-assignments using the tagged-template-string syntax of es6.
DSL-identifiers are configured in the configuration key `Neos.Fusion.dsl`. The configured objects must satisfy the `DslInterface` and return fusion code that is parsed by the fusion-parser afterwards.

```
value = dslExample`... the code that is passed to the dsl ...`
```
In addition this pr adds a schema for the fusion part of the Settings and integrates it into the automatic schema-validation.