Merge pull request #1476 from CORE4/fallbackgraph-visualization
FEATURE: Fallback graph visualization

The content repository is extended by two essential features,

* The Intra Dimensional Fallback Graph
* The Inter Dimensional Fallback Graph

which are supposed to be used for graph-aware projections in future versions.

These can be populated in-memory from the registered ``DimensionPresetSourceInterface`` by an application service and thus provide a read-only interface for applications in need of fallback logic.

In addition, Neos is extended by a backend module that visualizes these fallbacks.
They are displayed as an interactive graph using SVG and vanilla JS.

This can be tested/verified by setting up an arbitrary dimension configuration and visiting the Dimensions administration module.