Merge pull request #1597 from davidspiola/improve-asset-replacement
TASK: Keep asset adjustments if asset size did not change

This adds a condition to compare the size of the old and new asset in order to verify
the size did not change. In this case the method refit() will not be executed to keep
asset adjustments. This is very handy if you upload/replace a image with enriched
meta data (eg. IPTC or Exif.)