Merge pull request #1102 from kitsunet/bugfix/1055-resource-publish-23
BUGFIX: FileSystemTarget publishes resources in public root

Due to the defined resource paths (which are not taken into account
for the FileSystemSymlinkTarget) only files in subdirectories of
the public resource path were published. This change adds the public
path itself and adjusts the PackageStorage to take into account
an empty relative publication path. Additionally does a slight
refactoring to make the whole class easier to follow.

Fixes: #1055