Merge pull request #1711 from grebaldi/bugfix/1710/discard-error
BUGFIX: Make sure nodes that are to be discarded are actually publishable

`TYPO3\Neos\Service\PublishingService` traverses all children of a node to discard that node recursively. Not all those nodes are actually publishable though, so the following batch discard operation fails with the Exception `#1395841899: Nodes in a in a workspace without a base workspace cannot be discarded`.

So, I added a filter to `TYPO3\Neos\Service\PublishingService::discardNode` that removes all nodes from the to-be-discarded list, that are in a different workspace than the given node.

Resolves: #1710