Merge pull request #1825 from daniellienert/bugfix/load-thumbnails-async
BUGFIX: Load thumbs in media browser asynchronous

When you open the media module with a lot of high-res images while the thumbnails are not already generated and cached, loading can take a long time without noticable feedback.

The templates use a setting not available since the media browser was split into it's own package. Instead the setting needs to be read from the Media package.