Merge pull request #1247 from kitsunet/task/remove-deprecated-base-vh
!!! TASK: Remove deprecated ``BaseViewhelper``

The BaseViewHelper is deprecated since 2 major versions and
it's not best practice to use base tags anymore. The functionality
can be replicated by other means if needed.

This is breaking if you use the ``BaseViewHelper`` in your templates.
You should replace that with a simple HTML ``<base>`` tag and get
the baseUri from request yourself.