Merge pull request #1964 from kitsunet/bugfix/dynamic-cache-tags-23
BUGFIX: Dynamic cache entries should be evaluated via runtime

Dynamic cache entries are complex and need a lot of information
to be evaluated correctly, therefore the runtime should always take
care of this. This is potentially a bit slower than the original
code but definitely more correct in outcome. Before tags were
definitely wrong as they were based on the parent elements
configuration.