Merge pull request #1266 from neos/skurfuerst-cached-configuration-bug
BUGFIX: if cached configuration does not contain all configuration types, ensure on next run the configuration is cached again

## Problem Description

the following happens in Production context, if letting the CLI compile request run through, but abort the web request.

In this case:
- the cached Configurations.php object does *not* contain information for NodeTypes etc.
- thus, node types are loaded on every request
- this leads to a slow system which can only be fixed by removing the cached Configuration file.

To solve this issue, we now check whether the unprocessed configuration contains the right infos before serializing them.