Merge pull request #991 from albe/scalar-typehints-cache
!!!TASK: Add PHP 7.0 scalar type hints to method arguments and return types

This is breaking because it changes the cache interfaces and all implementations need to add scalar type hints.
The scope should be very limited though, since the amount of custom cache implementations is probably low. The most likely break will be with custom `*Privilege` implementations that also override the `getCacheEntryIdentifier()` method.