Merge pull request #415 from albe/distinct_implicit_joins
!!!TASK: Implicit joins in a query makes the result set distinct by default

Before it was possible that a query with implicit joins created a result set with duplicate entries.
This change sets the `distinct` flag on the query object to true when an implicit join is created.

This is breaking, as it might change the result sets of your queries, if you used implicit joins (dot notation subproperty queries) and you don't expect results to be distinct.

Resolves #552