Merge pull request #1273 from albe/safe-trusted-proxy-defaults
!!!TASK: Safe trusted proxies default value

By default, all proxies were trusted beforehand, but this is an usafe setting in most setups.
This change switches the `trustedProxies.proxies` setting to `%env:FLOW_HTTP_TRUSTED_PROXIES`, which means no proxies are trusted by default unless something is specified via the environment variable and hence the client IP address, port and host values for the request can not be overridden by any of the `Forwarded-*` headers.

This is breaking if you use a CDN or reverse proxy on your server and relied on the previous unsafe
default setting. In that case you should instead provide a list of the IP ranges of your proxy/CDN
servers, either directly or through the `FLOW_HTTP_TRUSTED_PROXIES` environment variable or explicitly switch back to trust all proxy servers by setting the value to `'*'`.