Merge pull request #1270 from albe/client-ip-address
TASK: Make HTTP Request getClientIpAddress() use the request attribute value

This in turn will actually make the method take the trusted proxy settings into account instead
of a hardcoded list of headers.
Also, this change removes the test for the `Forwarded` header,
because the test had wrong expectations on the actual formatting of the header value.
See #1060