Merge pull request #615 from gerhard-boden/update-root-typoscript-master
!!! FEATURE: Use prototypes instead of the ``page`` path

A freshly created site package will now create a fusion protoype that extends `Neos.Neos:Page`. Using the Fusion path `/page` can be seen as deprecated and packages that rely on extending `page` should be updated to extend the page prototype instead.

As a workaround the line `page = Neos.NodeTypes:Page` can be added to the `Root.fusion` as a fallback so extension points using the the `page` path as well as layout properties will still be in effect.

For further information see: http://neos.readthedocs.io/en/stable/HowTos/RenderingCustomDocumentTypes.html