Merge pull request #1276 from hlubek/feature-eel-arrow-function
FEATURE: Arrow function expressions in Eel

Adds support to parse and evaluate arrow function expressions in Eel.
This is only implemented for the CompilingEelEvaluator, since
expression evaluation need to be deferred in functions.

Arrow function arguments can be specified with or without parentheses.
The body must be an expression and not a block.

Example:

    map(items, (x) => x * x)