Merge pull request #1377 from kdambekalns/task/identityroutepart-lookup-speedup
TASK: Speed up lookups in IdentityRoutePart

This adds a second index to the database table in which the
ObjectPathMapping instances are stored. It speeds up find operations
for which the existing primary key does not help.

With enough data (250k entries for ObjectPathMapping) this speeds up
page delivery by about 60%.