!!! BUGFIX: Respect disabled superTypes when filtering by Node Type (#2147)
This fixes `NodeType::isOfType()` to return `false` when the given
type is explicitly disabled by the corresponding NodeType or one of
it's super types.

Example:

    'Acme:Animal':
      superTypes:
        'Acme:HasBonesMixin': true

    'Acme:Jellyfish':
      superTypes:
        'Acme:Animal': true
        'Acme:HasBonesMixin': false

With this `NodeTypeManager::getNodeType('Acme.Jellyfish')->isOfType('Acme:HasBonesMixin')`
should return `false` but it didn't.
Respectively a FlowQuery like the following should not return any "Jellyfish"-node (or
nodes with a sub-type): `q(node).find('[instanceof Acme:HasBonesMixin]')` but it did.

This might be a breaking change if Fusion or PHP code relies on the broken behavior.
NodeTypes with disabled `superTypes` should be tested for this.

Fixes: #1983