Merge pull request #2192 from mficzel/feature/fusionSpreads
FEATURE: Add ``@apply`` meta-keyword to the Fusion-language

``@apply`` allows to override multiple properties of a fusion-prototype with a single expression. This is useful when complex data structures are mapped to fusion prototypes.

The example shows the rendering of a ``teaserList``-array by using a Teaser-Component and passing all keys from each teaser to the fusion Object.

```
teasers = Neos.Fusion:Collection {
	collection = ${teaserList}
	itemName = 'teaser'
	itemRenderer = Vendor.Site:Teaser {
		@apply.teasers = ${teaser}
	}
}
```

The code avoids passing down each fusion-property explicitly to the child component. A similar concept with different syntax from the JavaScript world is known as ES6-Spreads.

Another use-case is to use ``Neos.Fusion:Renderer`` to render a prototype while type and properties are based on data from the context.

```
example = Neos.Fusion:Renderer {
	type = ${data.type}
	element.@apply.properties = ${data.properties}
}
```

That way some meta-programming can used in fusion and both prototype and properties are decided late in the rendering by the fusion runtime.

The keys below ``@apply`` are evaluated before the fusion-object and the ``@context`` and are initialized. Each key below ``@apply`` must return a key-value map (values other than an array it are ignored). During the evaluation of each fusion-path the values from ``@apply`` are always checked first.

If a property is defined via ``@apply`` this value is returned without evaluating the fusionPath.

The ``@process`` and ``@if``-rules of the original fusion-key are still applied even if a value from ``@apply`` is returned.

Since ``@apply`` is evaluated first the overwritten values are already present during the evaluation of ``@context`` and will overlay the properties of ``this`` if they are accessed.

``@apply`` supports the same extended syntax and ordering as fusion processors and supports multiple keys. The evaluation order is defined via ``@position``, the keys that are evaluated last will override previously defined keys. This is also similar to the rules for ``@process``.

```
test = Vendor.Site:Prototype {
	@apply.contextValue {
		@position = 'start'
		expression = ${ arrayValueFromContext }
	}
	@apply.fusionObject {
		@position = 'end'
		expression = Neos.Fusion:RawArray {
			value = "altered value"
		}
	}
}
```

Other than ``@context`` ``@apply`` is only valid for a single fusion path, so when subpathes or children are rendered they are not affected by the parents ``@apply`` unless they are explicitly passed down.