Merge pull request #2189 from mficzel/feature/replaceConfusingRawFusionPrototypes
FEATURE: Add Fusion prototypes `Join`, `Loop`, `Map`, `Reduce` and `DataStructure`

The difference between ``RawArray``, ``Array``, ``RawCollection`` and ``Collection`` was often hard to get for new developers. To overcome this the old confusing names are deprecated and new prototypes are introduced wich are easier to get and emphasize the declarative nature of Fusion.

``Neos.Fusion:Join``

The prototype concatenates the fusion values of all fusion-properties and returns the result as a string. This prototype replaces ``Neos.Fusion:Array`` wich is deprecated.

In addition to the behavior of the ``Array`` ``Join`` allows to define the ``@glue`` used for concatenating the parts.

``Neos.Fusion:Loop``

The ``Loop`` prototype iterates over the given items with the itemRenderer and returns the concatenated result as a string. The prototype replaces ``Neos.Fusion:Collection`` wich is deprecated.

Other than in ``Collection`` the items are passed with the key ``items`` instead of ``collection``.

In addition to the behavior of the classic ``Collection`` ``Loop ``  allows to define the ``@glue`` used for concatenating the items.

``Neos.Fusion:DataStructure``

The ``DataStructure`` prototype returns an associative array with all fusion keys evaluated. The prototype replaces ``Neos.Fusion:RawArray`` wich is deprecated.

``Neos.Fusion:Map``

The ``Map`` iterates over the given ``items`` and returns the result as array.  The prototype replaces ``Neos.Fusion:RawCollection`` wich is deprecated.

Other than ``RawCollection`` the items are passed with the key ``items`` instead of ``collection`` and the keys of the given ``items`` are preserved.

``Neos.Fusion:Reduce``

The `Neos.Fusion:Reduce` prototype is added wich reduces the given items to a single value by using ``itemRenderer`` with the following properties.

- `items` (array/Iterable, **required**) The array or iterable to iterate over
- `itemName`: (string, defaults to `item`) Context variable name for each item
- `itemKey`: (string, defaults to `itemKey`) Context variable name for each item key, when working with array
- `carryName`: (string, defaults to `carry`) Context variable that contains the result of the last iteration
- `iterationName`: (string, defaults to `iterator`) A context variable with iteration information will be available under the given name: ``index`` (zero-based), ``cycle`` (1-based), ``isFirst``, ``isLast``
- `itemReducer`: (mixed, **required**) The reducer definition (simple value, expression or object) that will be applied for every item.
- `initialValue`: (mixed, defaults to `null`) The value that is passed to the first iteration or returned if the items are empty