Merge pull request #2389 from mficzel/configureImageFormat
FEATURE: Add ``format`` option for image-prototypes, viewHelpers and presets

The option format is added to imagePresets, viewHelpers and FusionPrototypes.
The format is passed as string "jpg", "jpeg", "gif", "png", "wbmp", "xbm", "webp" and "bmp" are supported. If no format is given the crops will use the format of the original image.

This allows to to enforce rendering of crops in jpeg or png but also adds support for rendering
of webp-images as alternate sources.

Additionally this pr adds the ConvertImageThumbnailGenerator that allows to specify images that should be converted for online-presentation to an optimized target format. While the default imageFormat is specified in the settings this can still be overridden via format option on
the thumbnail-configurations.