Merge pull request #1549 from mficzel/bugfix/packageCreateWorksWithPrivatePackagist
BUGFIX: Fix package:create and derived commands when private packagist is used

When private packagist is used the following setting isn added to the repositories section of the composer.json:

```
repositories: [
    {
        "packagist.org": false
    }
]
```

This caused an error because the package:create command tried to access the undefined type property of each defined repository.

This change simply checks for the existence of the type key before accessing it.