Merge pull request #2489 from mficzel/task/improveErrorMessageForMissingFusionPrototypes
BUGFIX: Make exception for non renderable fusion path more helpful

The existing exception for non renderable fusion pathes was not very helpful hard to read and missed mentioning likely reasons like a typo in the prototype name. It also suggested a solution that is unlikely to fix the problem.

With this change the prototype name is moved to the front of the error message as it is the most important information.

The path is removed from the message-body as the fusion exception handler will render it anyway.

The most likely reasons typo + missing prototype are mentioned first and other possible reasons like missing parent-protopype, missing `@class` and missing `include:` later.

The proposed solution to inherit from `Neos.Fusion:Template` is removed as it was misleading